+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/napplet/protocol/NappletRequestSignerOp.kt
@@ -18,26 +18,21 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.commons.napplet.signers
+package com.vitorpamplona.amethyst.commons.napplet.protocol
+
+import com.vitorpamplona.amethyst.commons.connectedApps.signers.NostrSignerOp
 
 /**
- * The user's top-level trust decision for one app's access to the internal Nostr signer.
- * Set once on first connection; governs all future signing/encryption operations unless
- * overridden by a per-operation [NostrOpDecision] in the [NostrSignerPermissionLedger].
+ * Maps a [NappletRequest] to the [NostrSignerOp] it represents, or `null` if the request
+ * does not involve signing or encryption.
+ *
+ * This lives on the napplet side (not in the generic signer-permission layer) because it
+ * is the napplet protocol's own translation into the shared [NostrSignerOp] vocabulary.
  */
-enum class AppSignerPolicy {
-    /** Auto-approve all signing and encryption operations (except payments, which always prompt). */
-    FULL_TRUST,
-
-    /**
-     * Auto-approve the most common, harmless operations: additive public content events — short
-     * notes, reposts, reactions, picture posts, public chat, highlights, comments, and user status
-     * (see [NostrSignerPermissionLedger.REASONABLE_SIGN_KINDS]) — plus encryption; ask before
-     * anything that could spend money, overwrite account config, delete content, or reveal private
-     * data (including decryption). A reasonable default for most apps.
-     */
-    REASONABLE,
-
-    /** Prompt before every single signing or encryption operation. */
-    PARANOID,
-}
+fun NappletRequest.toSignerOp(): NostrSignerOp? =
+    when (this) {
+        is NappletRequest.Publish -> NostrSignerOp.SignKind(kind)
+        is NappletRequest.SignEvent -> NostrSignerOp.SignKind(kind)
+        is NappletRequest.PublishEncrypted -> NostrSignerOp.Encrypt
+        else -> null
+    }
